Reworked performance counter category creation code and associated tests so that counter category creation is part of EnablePerformanceCounters but can be supported as a deployment action.
